Corrected and sped up pulse rate calculation
Previous calculation was a ratio of pulses+1 to time elapsed. This was a
bug that resulted in inaccurate pulse rate calculation. Calculation has
been corrected to a ratio of pulses to time elapsed. In addition, pulse
rate calculation is performed every 3 pulses, rather than every 10. This
results in a faster response to pulse rate changes.